+++ Tests/Tests/Helpers/TestHelpers.swift
@@ -34,7 +34,8 @@ struct TestHelpers {
                                  hasImage: Bool = false,
                                  content: Content? = nil,
                                  triggers: [Trigger]? = nil,
-                                 buttons: [Button] = []) -> Campaign {
+                                 buttons: [Button] = [],
+                                 customJson: CustomJson? = nil) -> Campaign {
         Campaign(
             data: CampaignData(
                 campaignId: id,
@@ -69,7 +70,8 @@ struct TestHelpers {
                         controlSettings: ControlSettings(
                             buttons: buttons,
                             content: content))
-                )
+                ),
+                customJson: customJson
             )
         )
     }
@@ -120,7 +122,8 @@ struct TestHelpers {
                         controlSettings: ControlSettings(
                             buttons: [],
                             content: nil))
-                )
+                ),
+                customJson: nil
             )
         )
     }